combined util functions into one utils module for simpler import
