--- common/create_evaluation_df.py
@@ -1,7 +1,0 @@
-def create_evaluation_df(predictions, test_inputs, H, scaler):
-    eval_df = pd.DataFrame(predictions, columns=['t+'+str(t) for t in range(1, H+1)])
-    eval_df['timestamp'] = test_inputs.dataframe.index
-    eval_df = pd.melt(eval_df, id_vars='timestamp', value_name='prediction', var_name='h')
-    eval_df['actual'] = np.transpose(test_inputs['target']).ravel()
-    eval_df[['prediction', 'actual']] = scaler.inverse_transform(eval_df[['prediction', 'actual']])
-    return eval_df
