--- common/load_data.py
@@ -1,19 +1,0 @@
-def load_data():
-
-    # read GEFCom2014 load data
-
-    data_dir = 'data/'
-
-    energy = pd.read_csv(os.path.join(data_dir, 'energy.csv'), parse_dates=['timestamp'])
-
-    # Reindex the dataframe such that the dataframe has a record for every time point
-    # between the minimum and maximum timestamp in the time series. This helps to 
-    # identify missing time periods in the data (there are none in this dataset).
-
-    energy.index = energy['timestamp']
-    energy = energy.reindex(pd.date_range(min(energy['timestamp']),
-                                          max(energy['timestamp']),
-                                          freq='H'))
-    energy = energy.drop('timestamp', axis=1)
-
-    return energy
