--- common/mape.py
@@ -1,2 +1,0 @@
-def mape(predictions, actuals):
-    return ((predictions - actuals).abs() / actuals).mean()
